// src/model/Project.ts
import { FileHolder } from 'angular2-image-upload';
import { Story } from './Story';
import { Reward } from './Reward';
export class Project{

   Catagory: String;
   Title: String;
   FundingDuration: String;
   FundingGoal: number;
   ImageUrl: String;
   Loucation: String;
   ShortDescription: String;
   RewardList =[];
   ProjectStory: Story;
   TotalFuneded:number;
   ToalVoteCount:number;
  


  


}